pyproject.toml
[project]
name = "discord-mcp"
version = "0.1.0"
description = "Discord MCP server for reading and sending Discord messages"
readme = "README.md"
authors = [
    { name = "elyx", email = "elio@pascarelli.com" }
]
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.52.0",
    "mcp>=1.9.3",
    "python-dotenv>=1.1.0",
    "typing-extensions>=4.14.0",
]

[project.scripts]
discord-mcp = "discord_mcp:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pyright>=1.1.401",
    "ruff>=0.11.13",
]
